Extract templateutil models
