+++ web/template/templateutil/util.go
@@ -6,48 +6,6 @@ import(
   "treit/api/trello"
 )
 
-type Setup struct {
-  ClientStateAuthKey string
-  LinkForAuthKey string
-  ClientStateTrelloUsername string
-  ActionFinishSetup string
-}
-
-type TrelloBoards struct {
-  ClientStateUseTrelloBoardId string
-  ActionUseTrelloBoard string
-  Boards []TrelloBoard
-  ClientStateTrelloBoardName string
-}
-
-type TrelloBoard struct {
-  BoardId string
-  BoardName string
-}
-
-type TrelloLists struct {
-  ClientStateAddCardName string
-  ClientStateAddCardIdInListId string
-  ActionAddCard string
-  Lists []TrelloList
-  ActionDeleteCard string
-  BoardId string
-  ClientStateUpdateCardInListId string
-  ActionMoveCardToList string
-}
-
-type TrelloList struct {
-  ListId string
-  ListName string
-  ListCards []TrelloCard
-}
-
-type TrelloCard struct {
-  CardId string
-  CardName string
-  CardShortUrl string
-}
-
 func GenerateForSetup(clientStateAuthKey string, linkForAuthKey string, clientStateTrelloUserName string, actionFinishSetup string) (tmpl *template.Template, model Setup) {
   tmpl = templateWithName("setup.html")
   model = createSetupModel(clientStateAuthKey, linkForAuthKey, clientStateTrelloUserName, actionFinishSetup)
